/*clarifications and correcting one bug*/
As best as I can tell, based on directly testing this, only both output scalings need to be changed, and not the output range numbers as well. So the sentence should read "If you want to invert your PWM signal, the sign for both output scalings has to be changed. (O: -10000 -10000 0 -10000 10000)".