Remove duplicate code from the ubuntu build scripts.
All the scripts are installing at least the common
dependencies, but the shell code that installs the common
dependencies was copied in each one of the scripts. This is
error prone. If the common dependencies change, then one
has to go to each one of these scripts to do the changes.

This commit changes that behavior and make all the ubuntu
scripts depend on another script that is keeping the common
dependencies. Keep on reading for more details.

The logical thing to do is that the scripts `ubuntu_sim.sh`,
`ubuntu_sim_ros_gazebo.sh` and `ubuntu_sim_nuttx.sh` should
depend on another script that is responsible to install the
common dependencies, and install on top of the common deps
the additional components that needs to be installed for
each specific script. This script is now called
`ubuntu_sim_ros_gazebo.sh`.

How it works?
When a user downloads e.g. the script `ubuntu_sim_ros_gazebo.sh`,
the first thing this script does is to download and source the
script called `ubuntu_sim_common_deps.sh` before installing the
specific ROS/Gazebo stuff.

The script `ubuntu_sim_common_deps.sh` is now the only
script that we have to maintain when the common dependencies
change.

Signed-off-by: Vangelis Tasoulas <vangelis@tasoulas.net>