Windows: split up methods, revise, add Cygwin
The whole Windows instuctions section seemed quite outdated and cramped
onto one page to me. I can imagine it's confusing to read like that and
while linux instructions are already split up to multiple pages and
regularly updated.

I started a take on improving the Windows section, sorry that this the
first commit got that big.