build script: add astyle, remove vim-common
I think adding `astyle` to the deafult installation makes sense because
then the `make format` command can be used. On the other hand I don't
see how `vim-common` is a needed package. People using the editor will
have it anyways and beginners will probably not decide because it's installed
here.